[ALOY-737] Integrate xml2tss into style generation

+++ Alloy/commands/generate/generateUtils.js
@@ -2,6 +2,7 @@ var basePath = '../../';
 var path = require('path'),
 	fs = require('fs'),
 	wrench = require('wrench'),
+	xml2tss = require('xml2tss'),
 	alloyRoot = path.join(__dirname,'..','..'),
 	_ = require(basePath + 'lib/alloy/underscore')._,
 	U = require(basePath + 'utils'),
@@ -17,7 +18,9 @@ function pad(x) {
 
 exports.generateMigrationFileName = function(t) {
 	var d = new Date;
-	var s = String(d.getUTCFullYear()) + String(pad(d.getUTCMonth())) + String(pad(d.getUTCDate())) + String(pad(d.getUTCHours())) + String(pad(d.getUTCMinutes())) + String(d.getUTCMilliseconds())
+	var s = String(d.getUTCFullYear()) + String(pad(d.getUTCMonth())) + 
+		String(pad(d.getUTCDate())) + String(pad(d.getUTCHours())) + 
+		String(pad(d.getUTCMinutes())) + String(d.getUTCMilliseconds());
 	return s + '_' + t;
 }
 
@@ -33,7 +36,8 @@ exports.generate = function(name, type, program, args) {
 		if (_.contains(['VIEW','CONTROLLER','STYLE'],type)) {
 			dir = path.join(dir,program.platform);
 		} else {
-			logger.warn('platform "' + program.platform + '" ignored, not used with type "' + type + '"');
+			logger.warn('platform "' + program.platform + 
+				'" ignored, not used with type "' + type + '"');
 		}
 	}
 
@@ -41,7 +45,7 @@ exports.generate = function(name, type, program, args) {
 	var file = path.join(dir,name + ext);
 
 	// see if the file already exists
-	if (path.existsSync(file) && !program.force) {
+	if (path.existsSync(file) && !program.force && type !== "STYLE") {
 		U.die(" file already exists: " + file);
 	}
 
@@ -51,15 +55,30 @@ exports.generate = function(name, type, program, args) {
 		wrench.mkdirSyncRecursive(fullDir);
 	}
 
-	// write the file out based on the given template
-	var templateContents = fs.readFileSync(templatePath,'utf8');
-	if (args.templateFunc) { templateContents = args.templateFunc(templateContents); }
-	var code = _.template(templateContents, args.template || {});
-	fs.writeFileSync(file, code);
+	// only use xml2tss to generate style if the partner view exists
+	var viewFile = path.join(paths.app, CONST.DIR['VIEW'], name + "." + 
+		CONST.FILE_EXT['VIEW']);
+	if (type === "STYLE" && path.existsSync(viewFile)) {
+		xml2tss.updateFile(viewFile, file, function(err,ok) {
+			if (ok) {
+				logger.info('Generated style named ' + name);
+			} else {
+				logger.warn('Style named ' + name + ' already up-to-date');
+			}
+		});
+	} else {
+		// write the file out based on the given template
+		var templateContents = fs.readFileSync(templatePath,'utf8');
+		if (args.templateFunc) { 
+			templateContents = args.templateFunc(templateContents); 
+		}
+		var code = _.template(templateContents, args.template || {});
+		fs.writeFileSync(file, code);
 
-	return {
-		file: file,
-		dir: fullDir,
-		code: code
-	};
+		return {
+			file: file,
+			dir: fullDir,
+			code: code
+		};
+	}
 }
